+++ Transports/com.community.netcode.transport.facepunch/CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this package will be documented in this file. The format 
 
 ### Changed
 - Targets the Netcode for GameObjects 1.0.0 package.
+- Renamed namespaces from MLAPI to Netcode.
 
 ### Removed
 - Removed support for channels.
+- No longer send 1 byte of channel information in each message.
 
 ## 1.0.0
-First version of the Enet transport as a Unity package.
+First version of the Facepunch Transport as a Unity package.